BulletTimeStudio/Edge/BulletTimeEdge/YIAgentServer/src/modules/Algo/requirements.txt
numpy
opencv-python
pandas
scipy
pillow
matplotlib
tensorflow
